Migrate to Typescript
Also restructured the entire repo and fixed a few bugs thanks to
type-checking.

+++ src/commands/emoji.ts
@@ -1,16 +1,20 @@
-const Command = require('../command');
-const { send, config } = require('../util/util');
+import Command from '../command';
+import { send, config } from '../util/util';
 
 const MAX_EMOJIS = 20;
 
-module.exports = new Command(
+export default new Command(
 	'emoji',
 	'Sends an animated emoji. Only works with animated emoji.',
 	`emoji <animated emoji name> <optional number to repeat 1-${MAX_EMOJIS}>`,
 	'emoji yeet 5',
-	(message, args) => {
+	async (message, args) => {
+		if (!message.guild) {
+			return;
+		}
+
 		const argsWords = args.split(' ');
-		const animatedEmojis = message.guild.emojis.cache.filter((r) => r.animated);
+		const animatedEmojis = message.guild.emojis.cache.filter((r) => Boolean(r.animated));
 		const emoji = animatedEmojis.find((r) => r.name === argsWords[0]);
 		if (emoji) {
 			message.delete();
@@ -20,7 +24,7 @@ module.exports = new Command(
 			}
 			send(message.channel, emoji.toString().repeat(numEmojis));
 		} else {
-			const randomEmoji = animatedEmojis.random()?.name || '';
+			const randomEmoji = animatedEmojis.random()?.name ?? '';
 			const emojiString = animatedEmojis.map((r) => r.toString()).join(' ');
 			send(message.channel, `Animated emojis you can use are: ${emojiString}. Try \`${config.get('options.prefix')}emoji ${randomEmoji}\``);
 		}
